# perf(fixer): enhance auto fix