# Description

Improve performance and memory usage in Fixer

```plain
old version:    9111        113605 ns/op        7812 B/op         138 allocs/op
new version:  10000        104953 ns/op        7317 B/op         133 allocs/op
```

- Execution time: 7.6% faster (104953 ns/op vs 113605 ns/op)
- Memory allocation: 6.3% reduction (7317 B/op vs 7812 B/op)
- Allocation count: 3.6% fewer allocations (133 allocs/op vs 138 allocs/op)